refactor(catalog): one home for products and services, tax rates included
/app/catalog and /app/dispatch/settings/products rendered the same
ProductsList, ProductEditor, GroupsList and GroupEditor, so the two surfaces
duplicated about 90 lines of identical draft and selection orchestration and
gave a merchant two places to look. The dispatch settings sidebar even labelled
its entry "Catalog" too, so the same word appeared twice in the nav pointing at
near-identical pages.

/app/catalog is now the only one, and it gains the tax-rates tab the settings
page owned. Tax rates are an org setting (documents.taxRates), which is the
argument for leaving them in settings, but that argument does not survive the
settings page's other two tabs being the same catalog lists that already live
at top level. One tab away from the prices they apply to beats a second
destination.

products-services-page.tsx is deleted (271 lines). The route stays as a
redirect rather than a deletion: it was linked from the dispatch
getting-started cards and the line-builder's catalog picker for months, so
bookmarks exist. It forwards the tab param and maps the one value that changed,
products to items; tax-rates keeps its name so the existing ?s=tax-rates deep
link still lands on the right tab.

Renamed to "Products and Services" in the sidebar and the route breadcrumb. Note
this is the label migration 102 moved catalog_item away from, to free the word
for the product entity added in 101, so the sidebar now reads "Products"
(product records) above "Products and Services" (catalog items). Requested
deliberately. Tab labels stay "Catalog items" and "Catalog groups" since those
match the entity labels.

Also repointed: both dispatch getting-started cards, the catalog picker footer
link, part-pricing-card's CATALOG_SETTINGS_HREF, six doc comments naming the
deleted file as the draft owner, and the duplicate settings sidebar entry with
its now-unused Tags import.
