+++ apps/web/src/components/money/ui/catalog-page.tsx
@@ -5,7 +5,7 @@ import { FeatureKey, PermissionKey } from '@auxx/lib/permissions/client'
 import { MainPageContent } from '@auxx/ui/components/main-page'
 import { ResponsiveTabs } from '@auxx/ui/components/responsive-tabs'
 import { generateId } from '@auxx/utils'
-import { Boxes, Lock, Package } from 'lucide-react'
+import { Boxes, Lock, Package, Percent } from 'lucide-react'
 import { useQueryState } from 'nuqs'
 import { useState } from 'react'
 import { EmptyState } from '~/components/global/empty-state'
@@ -18,39 +18,52 @@ import { GroupEditor } from './settings/group-editor'
 import { GroupsList } from './settings/groups-list'
 import { ProductEditor } from './settings/product-editor'
 import { ProductsList } from './settings/products-list'
+import { TaxRateEditor } from './settings/tax-rate-editor'
+import type { TaxRate } from './settings/tax-rate-types'
+import { TaxRatesList } from './settings/tax-rates-list'
 
-type CatalogTab = 'items' | 'groups'
+type CatalogTab = 'items' | 'groups' | 'tax-rates'
 
 const TABS: { value: CatalogTab; label: string; icon: typeof Package }[] = [
   { value: 'items', label: 'Catalog items', icon: Package },
   { value: 'groups', label: 'Catalog groups', icon: Boxes },
+  { value: 'tax-rates', label: 'Tax rates', icon: Percent },
 ]
 
 /**
- * First-class catalog surface at `/app/catalog`
- * (plans/products/01-product-family.md §6 — surface promotion).
+ * Products and Services, the single home for the sellable catalog, at
+ * `/app/catalog` (plans/products/01-product-family.md §6, surface promotion).
  *
- * Hosts the SAME list/editor components as the dispatch settings tab at
- * `/app/dispatch/settings/products` — `ProductsList`/`ProductEditor` and
- * `GroupsList`/`GroupEditor` — inside the `/app/products` route shell instead
- * of the settings chrome. The settings tab keeps working unchanged (it also
- * owns tax rates, which stay a settings concern); `catalog_item` and
- * `catalog_group` stay `isVisible: false`, so this route plus its deliberate
- * sidebar entry IS the promotion — not a visibility flip.
+ * This route is now the ONLY one. `/app/dispatch/settings/products` used to
+ * render an identical copy of these same lists and editors plus the tax-rates
+ * tab, so the two surfaces duplicated ~90 lines of draft/selection
+ * orchestration and gave a merchant two places to look; that route is now a
+ * redirect here (tab query preserved) and its dispatch-settings sidebar entry
+ * is gone. Tax rates moved with it, they are an org setting
+ * (`documents.taxRates`), but keeping them one tab away from the prices they
+ * apply to beats keeping them beside a settings page that no longer exists.
  *
- * The selection/draft orchestration mirrors `settings/products-services-page.tsx`
- * (money 15-settings-phantom-editors.md phase 2): one phantom draft per tab,
- * dropped when untouched on selecting another row or switching tabs.
+ * `catalog_item` and `catalog_group` stay `isVisible: false`, so the entity
+ * sidebar never auto-links them, this route plus its deliberate sidebar entry
+ * IS the promotion, not a visibility flip.
+ *
+ * One phantom draft per record tab (money 15-settings-phantom-editors.md phase
+ * 2), dropped when untouched on selecting another row or switching tabs. Tax
+ * rates have no draft: they are a settings array, committed on add.
  */
 export function CatalogPage() {
   useRequireCapability(PermissionKey.settingsManage)
   const { hasAccess } = useFeatureFlags()
 
+  // `tax-rates` is the value the old settings route used, so the deep links
+  // that carried `?s=tax-rates` there keep landing on this tab.
   const [tab, setTab] = useQueryState('s', { defaultValue: 'items' as string })
-  const activeTab: CatalogTab = tab === 'groups' ? 'groups' : 'items'
+  const activeTab: CatalogTab =
+    tab === 'groups' ? 'groups' : tab === 'tax-rates' ? 'tax-rates' : 'items'
 
   const [selectedItemId, setSelectedItemId] = useState<string | null>(null)
   const [selectedGroupId, setSelectedGroupId] = useState<string | null>(null)
+  const [selectedTaxRateId, setSelectedTaxRateId] = useState<string | null>(null)
   const [itemDraft, setItemDraft] = useState<CatalogDraftHandle | null>(null)
   const [groupDraft, setGroupDraft] = useState<CatalogDraftHandle | null>(null)
 
@@ -106,15 +119,23 @@ export function CatalogPage() {
     if (groupDraft) setGroupDraft(null)
   }
 
-  const { getSetting } = useSettings({ scope: 'GENERAL' })
-  const currency = (getSetting('organization.currency') as string) || 'USD'
+  // `useSettings({ scope })` FILTERS reads to that scope (use-settings.tsx:44-54), currency
+  // stayed GENERAL while taxRates moved to DOCUMENTS (money MQ2 §A.3), so two hook instances
+  // are needed; each scope's `updateOrganizationSetting` still writes any key correctly
+  // (the mutation isn't scope-gated), so either is fine to use for tax-rate writes.
+  const { getSetting: getGeneralSetting } = useSettings({ scope: 'GENERAL' })
+  const { getSetting: getDocumentsSetting, updateOrganizationSetting } = useSettings({
+    scope: 'DOCUMENTS',
+  })
+  const currency = (getGeneralSetting('organization.currency') as string) || 'USD'
+  const taxRates = (getDocumentsSetting('documents.taxRates') as TaxRate[] | null) ?? []
 
   if (!hasAccess(FeatureKey.dispatch)) {
     return (
       <MainPageContent>
         <EmptyState
           icon={Lock}
-          title='Catalog Not Available'
+          title='Products and Services Not Available'
           description='Upgrade your plan to use quoting and dispatch.'
           button={<div className='h-12' />}
         />
@@ -122,7 +143,47 @@ export function CatalogPage() {
     )
   }
 
-  const selectedId = activeTab === 'items' ? selectedItemId : selectedGroupId
+  function commitTaxRates(next: TaxRate[]) {
+    updateOrganizationSetting('documents.taxRates', next)
+  }
+
+  function handleAddTaxRate() {
+    const id = generateId('taxrate')
+    const next: TaxRate[] = [
+      ...taxRates,
+      { id, name: 'New tax rate', rate: 0, isDefault: taxRates.length === 0 },
+    ]
+    commitTaxRates(next)
+    setSelectedTaxRateId(id)
+  }
+
+  function handleUpdateTaxRate(id: string, patch: Partial<TaxRate>) {
+    commitTaxRates(taxRates.map((r) => (r.id === id ? { ...r, ...patch } : r)))
+  }
+
+  function handleSetDefaultTaxRate(id: string) {
+    commitTaxRates(taxRates.map((r) => ({ ...r, isDefault: r.id === id })))
+  }
+
+  function handleDeleteTaxRate(id: string) {
+    const removed = taxRates.find((r) => r.id === id)
+    const next = taxRates.filter((r) => r.id !== id)
+    // Deleting the default promotes the first remaining rate, one is always default.
+    const first = next[0]
+    if (removed?.isDefault && first && !next.some((r) => r.isDefault)) {
+      next[0] = { ...first, isDefault: true }
+    }
+    commitTaxRates(next)
+    if (selectedTaxRateId === id) setSelectedTaxRateId(null)
+  }
+
+  const selectedTaxRate = taxRates.find((r) => r.id === selectedTaxRateId) ?? null
+  const selectedId =
+    activeTab === 'items'
+      ? selectedItemId
+      : activeTab === 'groups'
+        ? selectedGroupId
+        : selectedTaxRateId
 
   const editorContent =
     activeTab === 'items' ? (
@@ -132,13 +193,18 @@ export function CatalogPage() {
         onDraftNameChange={handleItemDraftNameChange}
         onDraftCommitted={handleItemDraftCommitted}
       />
-    ) : (
+    ) : activeTab === 'groups' ? (
       <GroupEditor
         selectedId={selectedGroupId}
         currency={currency}
         draft={groupDraft}
         onDraftNameChange={handleGroupDraftNameChange}
         onDraftCommitted={handleGroupDraftCommitted}
+      />
+    ) : (
+      <TaxRateEditor
+        taxRate={selectedTaxRate}
+        onUpdate={(patch) => selectedTaxRate && handleUpdateTaxRate(selectedTaxRate.id, patch)}
       />
     )
 
@@ -160,11 +226,18 @@ export function CatalogPage() {
           id='money-catalog'
           scroll='columns'
           pane={editorContent}
-          paneTitle={activeTab === 'items' ? 'Edit item' : 'Edit group'}
+          paneTitle={
+            activeTab === 'items'
+              ? 'Edit item'
+              : activeTab === 'groups'
+                ? 'Edit group'
+                : 'Edit tax rate'
+          }
           paneOpen={!!selectedId}
           onPaneClose={() => {
             setSelectedItemId(null)
             setSelectedGroupId(null)
+            setSelectedTaxRateId(null)
             setItemDraft(null)
             setGroupDraft(null)
           }}>
@@ -176,13 +249,22 @@ export function CatalogPage() {
               draft={itemDraft}
               onAddDraft={handleAddItemDraft}
             />
-          ) : (
+          ) : activeTab === 'groups' ? (
             <GroupsList
               selectedId={selectedGroupId}
               onSelect={handleSelectGroup}
               currency={currency}
               draft={groupDraft}
               onAddDraft={handleAddGroupDraft}
+            />
+          ) : (
+            <TaxRatesList
+              taxRates={taxRates}
+              selectedId={selectedTaxRateId}
+              onSelect={setSelectedTaxRateId}
+              onAdd={handleAddTaxRate}
+              onSetDefault={handleSetDefaultTaxRate}
+              onDelete={handleDeleteTaxRate}
             />
           )}
         </MasterDetailSplit>
